app/requirements.txt
asgiref==3.8.1
attrs==24.2.0
Django==5.1.3
django-cors-headers==4.6.0
django-filter==24.3
django-seed==0.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
Faker==33.1.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.21.0
six==1.16.0
sqlparse==0.5.2
toposort==1.10
typing_extensions==4.12.2
uritemplate==4.1.1
